requirements.txt
fastapi
uvicorn
streamlit
PyMuPDF
sentence-transformers
transformers
torch
numpy
faiss-cpu
langchain-google-genai
azure-cognitiveservices-speech 
pydub
